Add files via upload
Not finished. Controls are explained in the game itself and it's pretty resilient to user error.
Currently there is no win condition, but your bomb counter goes down for every flag you place, so if you have all tiles revealed and a bombcounter of 0 you 'win'.
I also intent to implement a feature where you can re-reveal a number to reveal all squares around that number that aren't flagged.